// src/@types/backend-api/types.gen.ts
export type PostMedicalRemindersSchedulesData = {
  requestBody?: {
    /**
     * ชื่อ
     */
    medicalName?: string;
    /**
     * สถานะการแจ้งเตือน
     */
    status?: string;
    /**
     * เขตเวลา
     */
    timezone?: string;
    /**
     * ประเภทการซ้ำ
     */
    repeatType?: "does_not_repeat" | "daily" | "weekly" | "monthly";
    /**
     * จำนวนโดส
     */
    dose?: number;
    /**
     * วันที่เริ่มต้น
     */
    startDate?: string;
    /**
     * วันที่สิ้นสุด
     */
    endDate?: string;
    /**
     * วันในสัปดาห์
     */
    dayOfWeeks?: Array<
      | "sunday"
      | "monday"
      | "tuesday"
      | "wednesday"
      | "thursday"
      | "friday"
      | "saturday"
    >;
    /**
     * วันที่ในเดือน
     */
    dateOfMonths?: Array<number>;
    /**
     * หน่วยของโดส
     */
    doseUnit?: string;
    /**
     * รายการเวลาการแจ้งเตือนทางการแพทย์
     */
    medicalReminderScheduleTimes?: Array<{
      /**
       * เวลาที่แจ้งเตือนทางการแพทย์
       */
      reminderTime: string;
      /**
       * เวลาอาหาร
       */
      mealTiming: "before_meal" | "after_meal" | "with_meal";
      /**
       * ช่วงเวลาอาหาร
       */
      mealPeriod: "breakfast" | "lunch" | "dinner" | "bedtime";
    }>;
    /**
     * รายการยาที่แนะนำ
     */
    recommendation?: Array<{
      /**
       * ชื่อยาที่แนะนำ
       */
      key: string;
      /**
       * ชื่อยาที่แนะนำ
       */
      label: string;
      /**
       * สถานะการใช้อินพุต
       */
      hasInput?: boolean;
      /**
       * ค่าอินพุต
       */
      inputValue?: string;
      /**
       * ค่าอินพุต
       */
      inputKey?: string;
      /**
       * หน่วยของยาที่แนะนำ
       */
      unit?: string;
    }>;
  };
};

export type PostMedicalRemindersSchedulesResponse = {
  /**
   * ไอดีของการแจ้งเตือนทางการแพทย์
   */
  id: string;
  /**
   * รายละเอียด
   */
  description?: string;
  /**
   * ไอดีของผู้ใช้
   */
  userId: string;
  /**
   * สถานะการแจ้งเตือน
   */
  status: string;
};

export type GetMedicalRemindersSchedulesByIdData = {
  id: string;
};

export type GetMedicalRemindersSchedulesByIdResponse = {
  item: {
    /**
     * ไอดีของการแจ้งเตือนทางการแพทย์
     */
    id: string;
    /**
     * ไอดีของผู้ใช้
     */
    userId?: string | null;
    /**
     * ชื่อ
     */
    medicalName: string;
    /**
     * วันและเวลาที่สร้าง
     */
    createdAt?: string;
    /**
     * วันและเวลาที่อัปเดตล่าสุด
     */
    updatedAt?: string;
    /**
     * วันและเวลาที่ลบ
     */
    deletedAt?: string | null;
    /**
     * ประเภทการซ้ำ
     */
    repeatType: string;
    /**
     * จำนวนโดส
     */
    dose: number;
    /**
     * วันที่เริ่มต้น
     */
    startDate?: string;
    /**
     * วันที่สิ้นสุด
     */
    endDate?: string;
    /**
     * สถานะการแจ้งเตือน
     */
    status: string;
    /**
     * วันในสัปดาห์
     */
    dayOfWeeks?: Array<
      | "sunday"
      | "monday"
      | "tuesday"
      | "wednesday"
      | "thursday"
      | "friday"
      | "saturday"
    >;
    /**
     * วันที่ในเดือน
     */
    dateOfMonths: Array<number>;
    /**
     * หน่วยของโดส
     */
    doseUnit: string;
    /**
     * เวลาอาหาร
     */
    mealTiming: string;
    /**
     * รายการยาที่แนะนำ
     */
    recommendation?: Array<{
      key: string;
      label: string;
      hasInput?: boolean;
      inputValue?: string;
      inputKey?: string;
      unit?: string;
    }>;
    medicalReminderScheduleTimes?: Array<{
      /**
       * ไอดีของการแจ้งเตือนทางการแพทย์
       */
      id: string;
      /**
       * ช่วงเวลาอาหาร
       */
      mealPeriod: string;
      /**
       * เวลาที่แจ้งเตือนทางการแพทย์
       */
      reminderTime: string;
    }>;
  };
};